norming
