+++ Makefile
@@ -6,7 +6,7 @@
 #    By: chlee2 <chlee2@student.hive.fi>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/01/06 14:27:00 by chlee2            #+#    #+#              #
-#    Updated: 2025/03/26 15:05:03 by chlee2           ###   ########.fr        #
+#    Updated: 2025/03/28 19:09:38 by mbutuzov         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -25,6 +25,7 @@ LDFLAGS = -L$(LIBFTDIR) -lft -lreadline
 SRCS = src/main.c \
 		src/parse.c \
 		src/sig.c \
+		src/sig_extra.c \
 		src/buildins/echo.c \
 		src/buildins/cd.c \
 		src/buildins/pwd.c \
@@ -35,20 +36,28 @@ SRCS = src/main.c \
 		src/lex/lex.c \
 		src/lex/wrong_pipe.c \
 		src/utils/free.c \
-		src/utils/print.c \
+		src/utils/free_extra.c \
 		src/lex/lex_helper.c \
 		src/pipex/command_ops.c \
 		src/pipex/error_ops.c \
+		src/pipex/error_ops_extra.c \
 		src/pipex/fd_ops.c \
+		src/pipex/fd_ops_extra.c \
+		src/pipex/fd_ops_extra_extra.c \
+		src/pipex/fd_ops_extra_extra_extra.c \
 		src/pipex/fork_ops.c \
+		src/pipex/fork_ops_extra.c \
 		src/pipex/parsing_utils.c \
 		src/pipex/pipex.c \
 		src/pipex/pipex_utils.c \
+		src/pipex/pipex_utils_extra.c \
 		src/pipex/utils.c \
 		src/pipex/heredoc.c \
+		src/pipex/heredoc_extra.c \
 		src/utils/lvl_handler.c \
 		src/utils/ft_split_join.c \
 		src/execute.c \
+		src/execute_extra.c \
 		src/ft_getenv.c \
 		src/lex/checkers.c \
 		src/lex/parse_input_character.c \
