+++ src/execute.c
@@ -6,7 +6,7 @@
 /*   By: chlee2 <chlee2@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/25 22:28:12 by mbutuzov          #+#    #+#             */
-/*   Updated: 2025/03/25 22:38:51 by mbutuzov         ###   ########.fr       */
+/*   Updated: 2025/03/28 18:55:17 by mbutuzov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,44 +72,4 @@ void	cleanup_parent_on_redirection_fail(t_shell *shell)
 	clear_cmds(shell);
 	shell->exit_code = 1;
 	restore_fds_and_error_exit(shell);
-}
-
-void	execute_cmd_parent(t_shell *shell)
-{
-	shell->cmds->shell = shell;
-	preserve_fds_and_error_exit(shell);
-	if (!resolve_heredoc_cmds(shell->cmds, 1))
-	{
-		if (shell->err_code)
-		{
-			shell->err_code = 0;
-			restore_fds_and_error_exit(shell);
-			return ;
-		}
-		else
-		{
-			print_error_message(error_init(HEREDOC_FAIL, 0, 0));
-			ft_free_all(shell);
-			exit(1);
-		}
-	}
-	if (process_file_redirections(shell->cmds))
-	{
-		cleanup_parent_on_redirection_fail(shell);
-		return ;
-	}
-	shell->exit_code = handle_builtin(*(shell->cmds));
-	restore_fds_and_error_exit(shell);
-}
-
-void	execute(t_shell *shell)
-{
-	if (!shell->cmds)
-	{
-		return ;
-	}
-	if (shell->cmds->next || !get_builtin_type(*(shell->cmds)))
-		shell->exit_code = pipex_launch(shell->cmds, shell->envp, shell);
-	else
-		execute_cmd_parent(shell);
 }
