+++ src/pipex/fork_ops.c
@@ -6,7 +6,7 @@
 /*   By: chlee2 <chlee2@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 22:28:08 by mbutuzov          #+#    #+#             */
-/*   Updated: 2025/03/25 18:45:00 by mbutuzov         ###   ########.fr       */
+/*   Updated: 2025/03/28 18:38:48 by mbutuzov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -104,42 +104,4 @@ int	after_fork(pid_t fork_result, t_pipex *pipex)
 	else if (pipex->current_command != pipex->command_count - 1)
 		ft_close(&(pipex->pipe[1]));
 	return (0);
-}
-
-int	get_exit_code(int wstatus)
-{
-	if (WIFEXITED(wstatus))
-		return (WEXITSTATUS(wstatus));
-	else if (WIFSIGNALED(wstatus))
-		return (128 + WTERMSIG(wstatus));
-	return (1);
-}
-
-int	wait_all(t_pipex pipex)
-{
-	int	wstatus;
-	int	exit_status;
-	int	sig_int_child;
-	int	wait_res;
-
-	sig_int_child = 0;
-	wstatus = 0;
-	exit_status = EXIT_FAILURE;
-	while (pipex.current_command--)
-	{
-		wait_res = wait(&wstatus);
-		if (wait_res == -1)
-			break ;
-		if (wait_res == pipex.last_pid)
-		{
-			exit_status = get_exit_code(wstatus);
-			if (exit_status == 130 && !sig_int_child++)
-				ft_putstr_fd("\n", 2);
-			if (exit_status == 131)
-				ft_putstr_fd("Quit (Core dumped)\n", 2);
-		}
-		else if (get_exit_code(wstatus) == 130 && !sig_int_child++)
-			ft_putstr_fd("\n", 2);
-	}
-	return (exit_status);
 }
