+++ src/sig.c
@@ -6,7 +6,7 @@
 /*   By: chlee2 <chlee2@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 19:16:25 by chlee2            #+#    #+#             */
-/*   Updated: 2025/03/26 22:33:11 by mbutuzov         ###   ########.fr       */
+/*   Updated: 2025/03/28 19:00:45 by mbutuzov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,30 +59,4 @@ void	before_child_process_signal(void)
 {
 	set_signal(SIGQUIT, SIG_IGN);
 	set_signal(SIGINT, SIG_IGN);
-}
-
-void	before_heredoc_process_signal(void)
-{
-	set_signal(SIGQUIT, SIG_IGN);
-	set_signal(SIGINT, SIG_IGN);
-}
-
-void	heredoc_int_beh(int asd)
-{
-	(void)asd;
-	g_sig = 1;
-	if (close(0) == -1)
-		perror("Can't close heredoc standard input");
-}
-
-void	set_child_signal(void)
-{
-	restore_signal(SIGQUIT);
-	restore_signal(SIGINT);
-}
-
-void	set_heredoc_signal(void)
-{
-	set_signal(SIGQUIT, SIG_IGN);
-	set_signal(SIGINT, heredoc_int_beh);
 }
